Implement unit tests for shorten endpoint

+++ src/main/scala/org/ottogiron/urlshortener/domain/urls/Url.scala
@@ -2,4 +2,4 @@ package org.ottogiron.urlshortener
 package domain.urls
 import java.time.Instant
 
-case class Url(hash: String, originalUrl: String, created: Instant, expirationDate: Instant, userId: Int)
+case class Url(hash: String, originalUrl: String, created: Instant, expirationDate: Instant, userId: Long)
